# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Anki card.",
        "title": "Anki card API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/card/generate": {
            "post": {
                "description": "Create  card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "Create card",
                "parameters": [
                    {
                        "description": "Create card",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateCard"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.CreateCard": {
            "type": "object",
            "required": [
                "material"
            ],
            "properties": {
                "apiKey": {
                    "type": "string"
                },
                "apiUrl": {
                    "type": "string"
                },
                "material": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}